+++ src/dashboardWebView/hooks/useMessages.tsx
@@ -15,7 +15,6 @@ import { Messenger } from '@estruyf/vscode/dist/client';
 import { EventData } from '@estruyf/vscode/dist/models';
 import { NavigationType } from '../models';
 import { GeneralCommands } from '../../constants';
-import * as l10n from '@vscode/l10n';
 
 export default function useMessages() {
   const [loading, setLoading] = useRecoilState(LoadingAtom);
@@ -25,7 +24,6 @@ export default function useMessages() {
   const [, setMode] = useRecoilState(ModeAtom);
   const [, setView] = useRecoilState(DashboardViewAtom);
   const [, setSearchReady] = useRecoilState(SearchReadyAtom);
-  const [localeReady, setLocaleReady] = useState<boolean>(false);
 
   const messageListener = (event: MessageEvent<EventData<any>>) => {
     const message = event.data;
@@ -61,12 +59,6 @@ export default function useMessages() {
       case GeneralCommands.toWebview.setMode:
         setMode(message.payload);
         break;
-      case GeneralCommands.toWebview.setLocalization:
-        l10n.config({
-          contents: message.payload
-        });
-        setLocaleReady(true);
-        break;
     }
   };
 
@@ -78,7 +70,6 @@ export default function useMessages() {
     Messenger.send(DashboardMessage.getTheme);
     Messenger.send(DashboardMessage.getData);
     Messenger.send(DashboardMessage.getMode);
-    Messenger.send(GeneralCommands.toVSCode.getLocalization);
 
     return () => {
       Messenger.unlisten(messageListener);
@@ -89,7 +80,6 @@ export default function useMessages() {
     loading,
     pages,
     viewData,
-    settings,
-    localeReady
+    settings
   };
 }