+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Change Log
 
-## [9.5.0] - 2024-xx-xx
+## [10.0.0] - 2024-xx-xx - [Release notes](https://beta.frontmatter.codes/updates/v10.0.0)
 
 ### ✨ New features
 
 - [#731](https://github.com/estruyf/vscode-front-matter/issues/731): Added the ability to map/unmap taxonomy to multiple pages at once
 - [#746](https://github.com/estruyf/vscode-front-matter/issues/746): Placeholder support added to to the `slug` field
 - [#749](https://github.com/estruyf/vscode-front-matter/issues/749): Ability to set your own filters on the content dashboard with the `frontMatter.content.filters` setting
+- [#756](https://github.com/estruyf/vscode-front-matter/issues/756): i18n/multilingual content support
 
 ### 🎨 Enhancements
 
@@ -17,8 +18,7 @@
 - [#741](https://github.com/estruyf/vscode-front-matter/issues/741): Added message on the content dashboard when content is processed
 - [#747](https://github.com/estruyf/vscode-front-matter/issues/747): The `@frontmatter/extensibility` dependency now supports scripts for placeholders
 - [#752](https://github.com/estruyf/vscode-front-matter/issues/752): Placeholder support in default `list` field values
-
-### ⚡️ Optimizations
+- Support for using the `fieldCollection` field in a `block` field
 
 - Updated the list of commands which are available in the command palette